requirements.txt
streamlit==1.30.0
Pillow==10.0.1
requests==2.31.0
